+++ tests/Feature/UniquePinnedPostTest.php
@@ -12,9 +12,7 @@ use Alley\WP\WP_Curate\Tests\TestCase;
 use function Mantle\Support\Helpers\html_string;
 
 /**
- * A test suite for unique pinned posts.
- *
- * @link https://mantle.alley.com/testing/test-framework.html
+ * Test the unique pinned post functionality.
  */
 class UniquePinnedPostTest extends TestCase {
 	/**
@@ -23,38 +21,9 @@ class UniquePinnedPostTest extends TestCase {
 	protected function setUp(): void {
 		parent::setUp();
 
-		$this->set_permalink_structure( '/%postname%/' );
-
-		static::factory()->post->create(
-			[
-				'post_title' => 'Test Post 1',
-			]
-		);
-		static::factory()->post->create(
-			[
-				'post_title' => 'Test Post 2',
-			]
-		);
-		static::factory()->post->create(
-			[
-				'post_title' => 'Test Post 3',
-			]
-		);
-		static::factory()->post->create(
-			[
-				'post_title' => 'Test Post 4',
-			]
-		);
-		static::factory()->post->create(
-			[
-				'post_title' => 'Test Post 5',
-			]
-		);
-		static::factory()->post->create(
-			[
-				'post_title' => 'Test Post 6',
-			]
-		);
+		for ( $i = 1; $i <= 6; $i++ ) {
+			static::factory()->post->create( [ 'post_title' => "Test Post $i" ] );
+		}
 	}
 
 	/**